+++ flake.nix
@@ -181,6 +181,13 @@
               ./nixos/staging.nix
             ];
           };
+          kpbj-stream-dev = nixpkgs.lib.nixosSystem {
+            system = "x86_64-linux";
+            modules = [
+              sops-nix.nixosModules.sops
+              ./nixos/dev.nix
+            ];
+          };
         };
     };
 }